requirements.txt
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.2
Flask==2.1.2
Flask-Login==0.6.1
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
oauthlib==3.2.0
pycparser==2.21
pyOpenSSL==22.0.0
python-dotenv==0.20.0
requests==2.27.1
urllib3==1.26.9
Werkzeug==2.1.2
